+++ src/chocolatey.ps1
@@ -14,6 +14,7 @@
   #[switch] $debug,
   [string] $name,
   [switch] $ignoreDependencies = $false,
+  [alias("x86")][switch] $forceX86 = $false,
   [parameter(Position=1, ValueFromRemainingArguments=$true)]
   [string[]]$packageNames=@('')
 )
@@ -113,7 +114,14 @@ if ([Environment]::OSVersion.Version -lt (new-object 'Version' 6,0)){
 #main entry point
 Append-Log
 
-Write-Debug "Arguments: `$command = '$command'|`$packageNames='$packageNames'|`$source='$source'|`$version='$version'|`$allVersions='$allVersions'|`$InstallArguments='$installArguments'|`$overrideArguments='$overrideArgs'|`$force='$force'|`$prerelease='$prerelease'|`$localonly='$localonly'|`$verbosity='$verbosity'|`$debug='$debug'|`$name='$name'|`$ignoreDependencies='$ignoreDependencies'"
+Write-Debug "Arguments: `$command = '$command'|`$packageNames='$packageNames'|`$source='$source'|`$version='$version'|`$allVersions=$allVersions|`$InstallArguments='$installArguments'|`$overrideArguments=$overrideArgs|`$force=$force|`$prerelease=$prerelease|`$localonly=$localonly|`$verbosity=$verbosity|`$debug=$debug|`$name='$name'|`$ignoreDependencies=$ignoreDependencies|`$forceX86=$forceX86"
+
+# run level environment variables
+$env:chocolateyForceX86 = $null
+if ($forceX86) {
+  $env:chocolateyForceX86 = $true
+}
+
 
 $chocolateyErrored = $false
 $badPackages = ''